Added a field to the MetricValue class that contains the unit of the metric.

+++ trunk/src/platform/toolkit/metrics/library/core/IMetered.cs
@@ -59,6 +59,6 @@ namespace Nohros.Metrics
     /// <summary>
     /// Gets the number of events which have been marked.
     /// </summary>
-    double Count { get; }
+    long Count { get; }
   }
 }
